refactor: modularize widget code into organized file structure
- Extract Provider, Entry, and Views into separate files
- Break down widget views into composable subviews
- Add medium font weight to status labels
